Добавил миграцию в таблицу users на поле id
- Эта правка связана с тем, что id'шни телеграм стали жирными
- Было Integer
- Стало Biginteger

+++ tables/user.py
@@ -5,4 +5,4 @@ from tables.base import Base
 
 class User(Base):
     __tablename__ = "users"
-    id = sa.Column(sa.Integer, primary_key=True, index=True)
+    id = sa.Column(sa.BigInteger, primary_key=True, index=True)
